refactor(api): move tag serializer
in src/organizational-entities
